backend/requirements.txt
apache-beam==2.61.0
apache-flink==2.0.0
apache-flink-libraries==2.0.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
avro==1.12.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==2.2.1
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==44.0.1
dill==0.3.1.1
dj-rest-auth==7.0.1
Django==4.2.19
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
dnspython==2.7.0
docopt==0.6.2
fastavro==1.10.0
fasteners==0.19
grpcio==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kazoo==2.5.0
MarkupSafe==3.0.2
numpy==2.1.3
oauthlib==3.2.2
objsize==0.7.1
openapi-codec==1.3.2
orjson==3.10.16
packaging==25.0
pandas==2.2.3
proto-plus==1.26.1
protobuf==5.29.4
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydot==1.4.2
PyJWT==2.10.1
pykafka==2.8.0
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
simplejson==3.20.1
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
zstandard==0.23.0
